FILE-STORAGE-REFACTOR: initial refactor
