+++ src/main/java/owpk/Application.java
@@ -3,9 +3,8 @@ package owpk;
 import io.micronaut.configuration.picocli.PicocliRunner;
 import lombok.extern.slf4j.Slf4j;
 import owpk.cli.ChatCommand;
-import owpk.storage.FileSettingsStore;
-import owpk.storage.main.MainSettingsStore;
-import owpk.storage.roles.RolesStorage;
+import owpk.storage.app.FileSettingsStore;
+import owpk.storage.app.MainSettingsStore;
 import owpk.utils.FileUtils;
 
 import java.io.IOException;
