+++ src/main/java/owpk/storage/ChatStorage.java
@@ -1,0 +1,4 @@
+package owpk.storage;
+
+public class ChatStorage {
+}
